add scenarios to set several vars at once
With named scenarios, we can encapsulate the 4 or more variables we would otherwise have to set.

Users will still be able to override those individual variables if they so desire.

I will make a follow up PR that will add an option in Jenkins to select from a drop down one of these scenarios.

I've also added a debug task to print out what values have been set based on the scenario.

We can still override all of this with setting the lower level variables.

For hop node scenarios, nodes are peered in a chain
automationcontroller -> hop1 -> hop2 ... -> hopN -> E1,E2,EN

Also tweaks the graph file param because when it is set to true,
it prints it to stdout instead of creating a file! :D

set defaults in the debug message

remove bad default, warn user early if they are missing config